% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{This is tidy data function}
\usage{
tidy_data(default_rates)
}
\arguments{
\item{default_rates}{uncleaned training data}
}
\description{
This function allows you to clean up and tidy the training data
required for student loan default prediction model.
}
\examples{
tidy_data()
}
\keyword{default_rate}
